+++ Delta.CertXplorer/src/plugins/PluralSightSelfCertPlugin/SelfSignedCertificatePlugin.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Drawing;
 using System.Windows.Forms;
-
 using Delta.CertXplorer.Extensibility;
-
+using Delta.CertXplorer.Extensibility.UI;
 using Pluralsight.Crypto.UI;
 
 namespace PluralSightSelfCertPlugin
@@ -46,7 +45,7 @@ namespace PluralSightSelfCertPlugin
                 base.Log.Error(ex);
                 var message = string.Format(
                     "There was an error while executing plugin {0}:\r\n\r\n{1}", PluginName, ex.Message);
-                MessageBox.Show(owner, message, "Error", MessageBoxButtons.OK);
+                ErrorBox.Show(owner, message);
 
                 return false;
             }